challenges/98-weather-agent/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "typer>=0.15.2",
    "genai-protocol>=1.0.9",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "nuitka>=2.6.9",
    "pyinstaller>=6.13.0",
]
